fix(gemma): use dtype instead of deprecated torch_dtype in from_pretrained
transformers 5.8.1 logs a warning_once for torch_dtype; switch to
the current kwarg name. Unit tests (test_vlm_slot.py) stay green.

+++ app/models/gemma_vlm.py
@@ -60,7 +60,7 @@ class GemmaVLM:
         self.model = AutoModelForMultimodalLM.from_pretrained(
             hf_repo,
             cache_dir=cache_dir,
-            torch_dtype=self.dtype,
+            dtype=self.dtype,
             device_map="auto" if self.device == "cuda" else None,
         )
         if self.device != "cuda":
